Se acomoda las estrucutras

+++ ProcessBase.py
@@ -19,9 +19,21 @@ class ProcessBase:
         with open(self.workspace_path +'/'+ self.temporal_parameterFile + ".xml", 'w') as f:
             f.write(self.temporal_parameterFile)
 
-    def find_files(self, file_extension):
-        return sorted(glob.glob(os.path.join(self.path_to_adq, f'*{file_extension}*')))
+    def find_files(self, file_extension, path_dttl=None):
+        # Si no se proporciona un path, usa self.path_to_adq
+        if path_dttl is None:
+            path_dttl = self.path_to_adq
+        return sorted(glob.glob(os.path.join(path_dttl, f'*{file_extension}*')))
 
     def move_files(self, files, destination_folder):
+        # Si files es una cadena de texto (un solo archivo), conviértelo en una lista
+        if isinstance(files, str):
+            files = [files]
+
         for v_file in files:
-            shutil.move(v_file, destination_folder)
+            print(f"Moviendo '{v_file}' a '{destination_folder}'")
+            # Verifica si el archivo existe antes de intentar moverlo
+            if os.path.isfile(v_file):
+                shutil.move(v_file, destination_folder)
+            else:
+                print(f"'{v_file}' no existe y no se moverá.")
